requirements.txt
web3
psycopg2-binary
sqlalchemy
pandas
python-dotenv
mysql-connector-python==8.0.25
requests==2.25.1
flask==1.1.2
